+++ openpub-base.php
@@ -5,9 +5,9 @@
  * Plugin URI:        https://www.openwebconcept.nl/
  * Description:       Acts as foundation for other OpenPub related content plugins. This plugin implements actions
  * to allow for other plugins to add and/or change Custom Posttypes, Metaboxes, Taxonomies, en Posts 2 posts relations.
- * Version:           1.1.4
- * Author:            Yard Internet
- * Author URI:        https://www.yardinternet.nl/
+ * Version:           1.1.5
+ * Author:            Yard Digital Agency
+ * Author URI:        https://www.yard.nl/
  * License:           GPL-3.0
  * License URI:       http://www.gnu.org/licenses/gpl-3.0.txt
  * Text Domain:       openpub-base