+++ src/Base/Foundation/Plugin.php
@@ -24,7 +24,7 @@ class Plugin
      *
      * @var string VERSION
      */
-    const VERSION = '1.1.2';
+    const VERSION = '1.1.5';
 
     /**
      * Path to the root of the plugin.