+++ src/Base/RestAPI/ItemFields/HighlightedItemField.php
@@ -24,9 +24,9 @@ class HighlightedItemField extends CreatesFields
      *
      * @param WP_Post $post
      *
-     * @return string
+     * @return bool
      */
-    private function getHighlightedItem(WP_Post $post)
+    private function getHighlightedItem(WP_Post $post): bool
     {
         return (bool) get_post_meta($post->ID, '_owc_openpub_highlighted_item', true) ?: false;
     }