+++ tests/Stubs/WordPress/WP_CLI.php
@@ -1,0 +1,8 @@
+<?php
+
+class WP_CLI
+{
+    public static function add_command()
+    {
+    }
+}