+++ tests/Unit/Config/config/directory/multi/deep.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-	'multi_level' => 'works'
-];