+++ tests/Unit/TestCase.php
@@ -7,20 +7,20 @@ use PHPUnit\Framework\TestCase as PHPUnit;
 class TestCase extends PHPUnit
 {
     /**
-	 * Call protected/private method of a class.
-	 *
-	 * @param object &$object    Instantiated object that we will run method on.
-	 * @param string $methodName Method name to call
-	 * @param array  $parameters Array of parameters to pass into method.
-	 *
-	 * @return mixed Method return.
-	 */
-	public function invokeMethod(&$object, $methodName, array $parameters = [])
-	{
-		$reflection = new \ReflectionClass(get_class($object));
-		$method     = $reflection->getMethod($methodName);
-		$method->setAccessible(true);
+     * Call protected/private method of a class.
+     *
+     * @param object &$object    Instantiated object that we will run method on.
+     * @param string $methodName Method name to call
+     * @param array  $parameters Array of parameters to pass into method.
+     *
+     * @return mixed Method return.
+     */
+    public function invokeMethod(&$object, string $methodName, array $parameters = [])
+    {
+        $reflection = new \ReflectionClass(get_class($object));
+        $method     = $reflection->getMethod($methodName);
+        $method->setAccessible(true);
 
-		return $method->invokeArgs($object, $parameters);
-	}
+        return $method->invokeArgs($object, $parameters);
+    }
 }